[management] split genesis tool from libra-management
In order to have a shared libary between multiple CLI tools, I've moved
the genesis tool to a separate crate.  This will let us have common code
between the tools.  Note: there are some rust files that are tied to the
genesis tool used in testing, that may not necessarily belong there as
the first step in this split.
