+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
     "config/generate-key",
     "config/global-constants",
     "config/management",
+    "config/management/genesis",
     "config/seed-peer-generator",
     "consensus",
     "consensus/consensus-types",
@@ -140,7 +141,7 @@ default-members = [
     "common/trace",
     "config/config-builder",
     "config/generate-key",
-    "config/management",
+    "config/management/genesis",
     "config/seed-peer-generator",
     "consensus/safety-rules",
     "client/libra-dev",
