+++ config/management/genesis/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "libra-genesis-tool"
+version = "0.1.0"
+authors = ["Libra Association <opensource@libra.org>"]
+description = "A tool to manage genesis"
+repository = "https://github.com/libra/libra"
+homepage = "https://libra.org"
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies]
+anyhow = "1.0.31"
+serde = { version = "1.0.114", features = ["rc"], default-features = false }
+structopt = "0.3.15"
+thiserror = "1.0.20"
+toml = { version = "0.5.6", default-features = false }
+
+config-builder = { path = "../../config-builder", version = "0.1.0" }
+executor = { path = "../../../execution/executor", version = "0.1.0" }
+generate-key = { path = "../../../config/generate-key", version = "0.1.0" }
+lcs = { path = "../../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
+libra-config = { path = "../..", version = "0.1.0"}
+libra-crypto = { path = "../../../crypto/crypto", version = "0.1.0" }
+libra-global-constants = { path = "../../../config/global-constants", version = "0.1.0" }
+libra-management = { path = "..", version = "0.1.0" }
+libra-network-address = { path = "../../../network/network-address", version = "0.1.0" }
+libra-secure-json-rpc = { path = "../../../secure/json-rpc", version = "0.1.0" }
+libra-secure-storage = { path = "../../../secure/storage", version = "0.1.0" }
+libra-secure-time = { path = "../../../secure/time", version = "0.1.0" }
+libra-types = { path = "../../../types", version = "0.1.0" }
+libra-workspace-hack = { path = "../../../common/workspace-hack", version = "0.1.0" }
+libra-temppath = { path = "../../../common/temppath", version = "0.1.0" }
+libra-vm = { path = "../../../language/libra-vm", version = "0.1.0" }
+libradb = { path = "../../../storage/libradb", version = "0.1.0" }
+storage-interface = { path = "../../../storage/storage-interface", version = "0.1.0" }
+transaction-builder = { path = "../../../language/transaction-builder", version = "0.1.0" }
+vm-genesis = { path = "../../../language/tools/vm-genesis", version = "0.1.0" }
+
+[features]
+testing = []
+fuzzing = ["libra-config/fuzzing"]
