+++ config/management/genesis/src/lib.rs
@@ -1,0 +1,11 @@
+// Copyright (c) The Libra Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+#![forbid(unsafe_code)]
+pub mod command;
+
+#[cfg(any(test, feature = "testing"))]
+mod storage_helper;
+
+#[cfg(any(test, feature = "testing"))]
+pub mod config_builder;
