+++ testsuite/Cargo.toml
@@ -25,6 +25,7 @@ generate-key = { path = "../config/generate-key", version = "0.1.0" }
 lcs = { path = "../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-config = { path = "../config", version = "0.1.0" }
 libra-crypto = { path = "../crypto/crypto", version = "0.1.0" }
+libra-genesis-tool = {path = "../config/management/genesis", version = "0.1.0", features = ["testing"] }
 libra-global-constants = { path = "../config/global-constants", version = "0.1.0" }
 libra-json-rpc = { path = "../json-rpc", version = "0.1.0" }
 libra-key-manager = { path = "../secure/key-manager", version = "0.1.0" }
