+++ testsuite/libra-swarm/Cargo.toml
@@ -21,6 +21,7 @@ generate-key = { path = "../../config/generate-key", version = "0.1.0" }
 libra-crypto = { path = "../../crypto/crypto", version = "0.1.0", features = ["cloneable-private-keys"] }
 libra-logger = { path = "../../common/logger", version = "0.1.0" }
 libra-management = {path = "../../config/management", version = "0.1.0", features = ["testing"] }
+libra-genesis-tool = {path = "../../config/management/genesis", version = "0.1.0", features = ["testing"] }
 libra-secure-storage = { path = "../../secure/storage", version = "0.1.0", features = ["testing"] }
 libra-temppath = { path = "../../common/temppath", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
