+++ testsuite/libra-swarm/src/swarm.rs
@@ -5,8 +5,8 @@ use anyhow::{Context, Result};
 use config_builder::SwarmConfig;
 use debug_interface::NodeDebugClient;
 use libra_config::config::{NodeConfig, RoleType};
+use libra_genesis_tool::config_builder::{FullnodeBuilder, FullnodeType, ValidatorBuilder};
 use libra_logger::prelude::*;
-use libra_management::config_builder::{FullnodeBuilder, FullnodeType, ValidatorBuilder};
 use libra_temppath::TempPath;
 use libra_types::account_address::AccountAddress;
 use std::{
